requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0,<2.0
django-ckeditor-5==0.2.19

djangorestframework
pytest
djangorestframework-simplejwt
django-cors-headers